requirements.txt
apify-client
pandas
python-dotenv
selenium
webdriver-manager